+++ assets/js/display/public.js
@@ -40,7 +40,6 @@ try {
     var easExtreme = false;
 
     // Define request object; this will be populated on socket connection
-    var hostReq;
     var noReq;
 
     // Define additional variables
@@ -205,33 +204,6 @@ try {
         html: `<h1 style="text-align: center; font-size: 3em; color: #FFFFFF">WWSU EAS - Active Alerts</h1><h2 style="text-align: center; font-size: 1.5em; color: #FFFFFF">Clark, Greene, and Montgomery counties</h2><div style="overflow-y: hidden;" class="d-flex flex-wrap" id="eas-alerts"></div>`,
     });
 
-    if (isStudio && moment().isBefore(moment("2019-02-12 00:00:00")))
-    {
-        // AAUP Strike
-        Slides.newSlide({
-            name: `aaup`,
-            label: `AAUP Strike`,
-            weight: -999999,
-            isSticky: true,
-            color: `success`,
-            active: true,
-            transitionIn: `fadeIn`,
-            transitionOut: `fadeOut`,
-            displayTime: 14,
-            fitContent: false,
-            html: ``,
-            reset: false
-        });
-
-        setInterval(() => {
-            Slides.slide(`aaup`).html = `
-            <h1 style="text-align: center; font-size: 3em; color: #FF7878">AAUP-WSU Strike</h1>
-            <div style="text-align: center; font-size: 6em; text-shadow: 4px 8px 6px rgba(0,0,0,0.3);" id="aaup-day" class="rounded bg-success text-white shadow-8">THE STRIKE IS OVER!!!</div>
-            <div style="text-align: center; font-size: 3em; color: #ffff78; text-shadow: 2px 4px 3px rgba(0,0,0,0.3);" id="aaup-total">Total time: 471 hours (20 days)</div>
-            <div style="text-align: center; font-size: 6em; text-shadow: 2px 4px 3px rgba(0,0,0,0.3);" class="m-3 text-success" id="aaup-strike2">Welcome back professors!</div>`;
-        }, 1000);
-
-    }
 
 // Create restart function to restart the screen after 15 seconds if it does not connect.
     var restart = setTimeout(function () {
@@ -1281,7 +1253,6 @@ waitFor(() => {
 }, () => {
 
     // Define hostReq object
-    hostReq = new WWSUreq(io.socket, `display-public`, 'host', '/auth/host', 'Host');
     noReq = new WWSUreq(io.socket, `display-public`);
 
     // When new Meta is received, update it in our memory and then run the process function.
@@ -1433,7 +1404,7 @@ waitFor(() => {
 function onlineSocket()
 {
     console.log('attempting online socket');
-    hostReq.request({method: 'POST', url: '/recipients/add-display', data: {host: 'display-public'}}, function (body) {
+    noReq.request({method: 'POST', url: '/recipients/add-display', data: {host: 'display-public'}}, function (body) {
         try {
         } catch (e) {
             console.log('FAILED ONLINE CONNECTION');